final servo

+++ servoMotor/Core/Src/servoMotor.h
@@ -14,16 +14,16 @@
 class servoMotor: public bandWithGenarator {
 public:
 	servoMotor();
-	servoMotor(TIM_HandleTypeDef*,float,float,double,double,int8_t,int8_t);
-	void gotoAngle(int8_t);
+	servoMotor(TIM_HandleTypeDef*,float,float,double,double,int8_t,int16_t);
+	void gotoAngle(int16_t);
 	int8_t getAngle(void);
 	virtual ~servoMotor();
 private:
 	float startBandWith;
 	float endBandWith;
 	int8_t angle;
-	std::function<double(int8_t)> mapAngle2BandWtih;
-	//auto mapAngle2BandWtih;
+	std::function<double(int16_t)> mapAngle2BandWtih;
+
 };
 
 #endif /* SRC_SERVOMOTOR_H_ */
